/macrodef/set/
