+++ src/rules/mod.rs
@@ -2,7 +2,7 @@ mod decision;
 pub use self::decision::Decision;
 
 mod file_type;
-pub use self::file_type::{MacroDef, Rvalue};
+pub use self::file_type::Rvalue;
 
 mod object;
 pub use self::object::Object;
@@ -17,6 +17,9 @@ pub use self::rule::Rule;
 
 mod subject;
 pub use self::subject::Subject;
+
+mod set;
+pub use self::set::Set;
 
 pub(crate) fn bool_to_c(b: bool) -> char {
     if b {
